final_project/VLM/requirements.txt
numpy
pandas
matplotlib
torch
torchvision
Pillow
transformers
qwen-vl-utils